feat: set creation time to the earliest of birthtime or modified time
I have files that exist on my Windows machine that somehow have a modified time that is earlier than the birthtime. I'm assuming that the earlier mtime is the true time the files were created, and that the birthtime was set at some later point when the files were copied.

Birthtime and modified time (mtime) work differently on Unix vs Windows. When a file is copied on Windows the birthtime of the original file is applied to the new file. On Unix the birthtime of the new file is set to the time at which the file is copied.

It shouldn't be the case then that the mtime is ever before the birthtime, no matter which OS, but here we are!