refactor: delete old helper method
The purpose of get_destination_directory() was to dynamically generate the expected destination directory of test files based on the creation date of the file. As time passes this date would change, making the tests fail. The new approach sets both the creation time of the test file and the destination directory path.